+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - ORM module
 
 ### [BUGFIXES]
-- none
+- [Issue #12](https://github.com/avolutions/avolutions/issues/12): Implemented exception handling for Config 
+- [Issue #14](https://github.com/avolutions/avolutions/issues/14): Parameter was not working correctly in Route
+- [Issue #15](https://github.com/avolutions/avolutions/issues/15): Optional parameter not working correctly in Route
 
 ## [2019-10-03] 0.4.0-alpha
 ### [FEATURES]